move get location to new client side file location.ts
