+++ frontend/app/lib/server/actions.ts
@@ -1,6 +1,6 @@
 "use server";
 
-import {getBackendURL, getLocationURL} from "./environment";
+import {getBackendURL} from "./environment";
 import {ChatCompletionMessageParam} from "@/app/domain/dashboard/messages/message.types";
 
 export async function queryFlights(messageHistory: ChatCompletionMessageParam[]) {
@@ -15,15 +15,6 @@ export async function queryFlights(messageHistory: ChatCompletionMessageParam[])
     if (!res.ok) {
         const errorMessage = await res.text();
         throw new Error(errorMessage || 'Could not connect to the server');
-    }
-    return await res.json();
-}
-
-//todo check if it gives the current location or the location from the frontend server
-export async function getUserLocation() {
-    const res = await fetch(`${await getLocationURL()}&fields=city,country_name,languages,currency`);
-    if (!res.ok) {
-        throw new Error('Could not connect to the server');
     }
     return await res.json();
 }
